2025046112-参赛总文件夹/2025046112-02素材与源码/requirements.txt
# Project Dependencies
# Python 3.12.2

numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tensorflow>=2.14.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
pyserial>=3.5
requests>=2.31.0

# If you have custom modules, ensure they are installed:
# kalman_people_flow_density
# simplified_sensor_simulator